# apis/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/patients": {
      "post": {
        "operationId": "PatientController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new patient",
        "tags": [
          "patients"
        ]
      },
      "get": {
        "operationId": "PatientController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all patients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all patients with optional filtering",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/patients/{id}": {
      "get": {
        "operationId": "PatientController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return patient details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get patient by ID",
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "PatientController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update patient by ID",
        "tags": [
          "patients"
        ]
      },
      "delete": {
        "operationId": "PatientController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient deleted successfully"
          },
          "404": {
            "description": "Patient not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete patient by ID",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/patients/{id}/photo": {
      "post": {
        "operationId": "PatientController_uploadPhoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload patient profile photo",
        "tags": [
          "patients"
        ]
      }
    }
  },
  "info": {
    "title": "KINE-SAAS API",
    "description": "API documentation for the KINE-SAAS physiotherapy backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "other",
          "prefer-not-to-say"
        ]
      },
      "MaritalStatus": {
        "type": "string",
        "enum": [
          "single",
          "married",
          "divorced",
          "widowed",
          "separated"
        ]
      },
      "PreferredContact": {
        "type": "string",
        "enum": [
          "phone",
          "email",
          "sms"
        ]
      },
      "CreateEmergencyContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "relationship": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "relationship",
          "phoneNumber"
        ]
      },
      "CreatePatientDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string"
          },
          "gender": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Gender"
              }
            ]
          },
          "maritalStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MaritalStatus"
              }
            ]
          },
          "phoneNumber": {
            "type": "string"
          },
          "alternativePhoneNumber": {
            "type": "string"
          },
          "preferredContact": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PreferredContact"
              }
            ]
          },
          "active": {
            "type": "boolean"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "kinesitherapeuteId": {
            "type": "string"
          },
          "emergencyContacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateEmergencyContactDto"
            }
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "dateOfBirth",
          "gender"
        ]
      },
      "MedicalRecordResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "medicalHistory": {
            "type": "string"
          },
          "allergies": {
            "type": "string"
          },
          "currentMedications": {
            "type": "string"
          },
          "previousTreatments": {
            "type": "string"
          },
          "diagnosisNotes": {
            "type": "string"
          },
          "treatmentPlan": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "InsuranceInfoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "policyNumber": {
            "type": "string"
          },
          "groupNumber": {
            "type": "string"
          },
          "primaryInsuredName": {
            "type": "string"
          },
          "relationship": {
            "type": "string"
          },
          "coverageDetails": {
            "type": "string"
          },
          "validUntil": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider",
          "policyNumber",
          "primaryInsuredName"
        ]
      },
      "EmergencyContactResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "relationship": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "relationship",
          "phoneNumber"
        ]
      },
      "PatientResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "profilePhotoUrl": {
            "type": "string"
          },
          "medicalRecord": {
            "$ref": "#/components/schemas/MedicalRecordResponseDto"
          },
          "insuranceInfo": {
            "$ref": "#/components/schemas/InsuranceInfoResponseDto"
          },
          "emergencyContacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmergencyContactResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "dateOfBirth",
          "gender",
          "phoneNumber",
          "address",
          "profilePhotoUrl",
          "medicalRecord",
          "insuranceInfo",
          "emergencyContacts",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePatientDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string"
          },
          "gender": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Gender"
              }
            ]
          },
          "maritalStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MaritalStatus"
              }
            ]
          },
          "phoneNumber": {
            "type": "string"
          },
          "alternativePhoneNumber": {
            "type": "string"
          },
          "preferredContact": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PreferredContact"
              }
            ]
          },
          "active": {
            "type": "boolean"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "kinesitherapeuteId": {
            "type": "string"
          },
          "emergencyContacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateEmergencyContactDto"
            }
          }
        }
      }
    }
  }
}